# Perf: Optimize vectorized append function